requirements.txt
streamlit
gspread
oauth2client
uuid
datetime
pandas
plotly.express
wordcloud